Updated db connections in student dashboard
