--- admin_data.csv
@@ -1,4 +1,0 @@
-Name,Employee Number,Date of Birth
-teacher1,12345,1/11/2001
-teacher2, 34567, 2/11/2001
-teacher3, 6789, 3/11/2001
