--- department_data.csv
@@ -1,2 +1,0 @@
-Name,Employee Number,Date of Birth
-teacher1,12345,1/11/2001
