--- student_data - Sheet1.csv
@@ -1,5 +1,0 @@
-Name,Register Number,Date of Birth
-Shreya Shukla,RA1911003010429,12/03/2000
-Anushka,RA1911003010450,23/02/2001
-Kushagra Gupta,RA1911003010389,12/06/2000
-Oishwarjya Banerjee,RA1911003010688,22/11/2001
